# backend/scratch/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIHire Recruitment Platform",
    "description": "Backend API service for the AI-powered Recruitment Platform.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/ml-prediction/job/{job_id}/candidate/{candidate_id}": {
      "post": {
        "tags": [
          "ML Suitability Prediction"
        ],
        "summary": "Get Ml Suitability Prediction",
        "description": "Runs the RandomForest ML suitability prediction for a candidate against a job profile.",
        "operationId": "get_ml_suitability_prediction_ml_prediction_job__job_id__candidate__candidate_id__post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MLPredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/assessment/start": {
      "post": {
        "tags": [
          "Aptitude Assessment"
        ],
        "summary": "Start or resume an aptitude assessment attempt",
        "operationId": "start_assessment_assessment_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentStartResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/assessment/result/{candidate_id}": {
      "get": {
        "tags": [
          "Aptitude Assessment"
        ],
        "summary": "Get latest assessment result for candidate",
        "operationId": "fetch_assessment_result_assessment_result__candidate_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assessment/reset": {
      "post": {
        "tags": [
          "Aptitude Assessment"
        ],
        "summary": "Reset candidate's assessment attempt and results to allow reattempt",
        "operationId": "reset_candidate_assessment_assessment_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/interview/start": {
      "post": {
        "tags": [
          "AI Interview Assessment"
        ],
        "summary": "Start or resume AI interview session",
        "operationId": "start_interview_interview_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewStartResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/interview/question": {
      "get": {
        "tags": [
          "AI Interview Assessment"
        ],
        "summary": "Get details of a specific question",
        "operationId": "get_question_interview_question_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "question_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewQuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interview/answer": {
      "post": {
        "tags": [
          "AI Interview Assessment"
        ],
        "summary": "Submit voice answer audio file",
        "operationId": "submit_answer_interview_answer_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_answer_interview_answer_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewAnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/interview/evaluate": {
      "post": {
        "tags": [
          "AI Interview Assessment"
        ],
        "summary": "Evaluate the submitted answer transcript",
        "operationId": "evaluate_answer_interview_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewEvaluateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/interview/finish": {
      "post": {
        "tags": [
          "AI Interview Assessment"
        ],
        "summary": "Finalize the interview session and compile report",
        "operationId": "finish_interview_interview_finish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interview/result/{candidate_id}": {
      "get": {
        "tags": [
          "AI Interview Assessment"
        ],
        "summary": "Get final interview assessment scorecard",
        "operationId": "get_result_interview_result__candidate_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interview/reset": {
      "post": {
        "tags": [
          "AI Interview Assessment"
        ],
        "summary": "Reset candidate's interview session and result to allow reattempt",
        "operationId": "reset_interview_interview_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentQuestionResponse": {
        "properties": {
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Options"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "selected_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Answer"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "question",
          "options",
          "category"
        ],
        "title": "AssessmentQuestionResponse"
      },
      "AssessmentResultResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "attempt_id": {
            "type": "integer",
            "title": "Attempt Id"
          },
          "aptitude_score": {
            "type": "number",
            "title": "Aptitude Score"
          },
          "quantitative_score": {
            "type": "number",
            "title": "Quantitative Score"
          },
          "logical_score": {
            "type": "number",
            "title": "Logical Score"
          },
          "verbal_score": {
            "type": "number",
            "title": "Verbal Score"
          },
          "analytical_reasoning_score": {
            "type": "number",
            "title": "Analytical Reasoning Score"
          },
          "computer_fundamentals_score": {
            "type": "number",
            "title": "Computer Fundamentals Score"
          },
          "total_correct": {
            "type": "integer",
            "title": "Total Correct"
          },
          "total_wrong": {
            "type": "integer",
            "title": "Total Wrong"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "candidate_id",
          "attempt_id",
          "aptitude_score",
          "quantitative_score",
          "logical_score",
          "verbal_score",
          "analytical_reasoning_score",
          "computer_fundamentals_score",
          "total_correct",
          "total_wrong",
          "status",
          "created_at"
        ],
        "title": "AssessmentResultResponse"
      },
...
      "AssessmentStartResponse": {
        "properties": {
          "attempt_id": {
            "type": "integer",
            "title": "Attempt Id"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "remaining_seconds": {
            "type": "integer",
            "title": "Remaining Seconds"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/AssessmentQuestionResponse"
            },
            "type": "array",
            "title": "Questions"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "duration",
          "remaining_seconds",
          "questions",
          "status"
        ],
        "title": "AssessmentStartResponse"
      },
...
      "Body_submit_answer_interview_answer_post": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "speech_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speech Text"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "question_id",
          "file"
        ],
        "title": "Body_submit_answer_interview_answer_post"
      },
...
      "EvaluateRequest": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "question_id"
        ],
        "title": "EvaluateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewAnswerResponse": {
        "properties": {
          "transcript": {
            "type": "string",
            "title": "Transcript"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "transcript",
          "question_id",
          "session_id"
        ],
        "title": "InterviewAnswerResponse"
      },
      "InterviewEvaluateResponse": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "communication_score": {
            "type": "number",
            "title": "Communication Score"
          },
          "technical_score": {
            "type": "number",
            "title": "Technical Score"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score"
          },
          "professionalism_score": {
            "type": "number",
            "title": "Professionalism Score"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "feedback": {
            "additionalProperties": true,
            "type": "object",
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "question_id",
          "communication_score",
          "technical_score",
          "confidence_score",
          "professionalism_score",
          "score",
          "feedback"
        ],
        "title": "InterviewEvaluateResponse"
      },
      "InterviewQuestionPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_text",
          "category",
          "order_index"
        ],
        "title": "InterviewQuestionPublic"
      },
      "InterviewQuestionResponse": {
        "properties": {
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "question_text",
          "category",
          "order_index",
          "total_questions"
        ],
        "title": "InterviewQuestionResponse"
      },
      "InterviewResultResponse": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "communication_score": {
            "type": "number",
            "title": "Communication Score"
          },
          "technical_score": {
            "type": "number",
            "title": "Technical Score"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score"
          },
          "professionalism_score": {
            "type": "number",
            "title": "Professionalism Score"
          },
          "total_score": {
            "type": "number",
            "title": "Total Score"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weaknesses"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "hiring_recommendation": {
            "type": "string",
            "title": "Hiring Recommendation"
          },
          "detailed_report": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailed Report"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "questions_answers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/QuestionAnswerItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questions Answers"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "session_id",
          "communication_score",
          "technical_score",
          "confidence_score",
          "professionalism_score",
          "total_score",
          "grade",
          "strengths",
          "weaknesses",
          "suggestions",
          "hiring_recommendation",
          "created_at"
        ],
        "title": "InterviewResultResponse"
      },
...
      "InterviewStartResponse": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/InterviewQuestionPublic"
            },
            "type": "array",
            "title": "Questions"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_index": {
            "type": "integer",
            "title": "Current Index"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "duration",
          "questions",
          "status",
          "current_index"
        ],
        "title": "InterviewStartResponse"
      },
...
      "MLPredictionResponse": {
        "properties": {
          "prediction": {
            "type": "string",
            "title": "Prediction"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "prediction",
          "probability",
          "confidence"
        ],
        "title": "MLPredictionResponse"
      },
...
      "QuestionAnswerItem": {
        "properties": {
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "transcript": {
            "type": "string",
            "title": "Transcript"
          }
        },
        "type": "object",
        "required": [
          "question_text",
          "category",
          "transcript"
        ],
        "title": "QuestionAnswerItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}